Added update and delete function

+++ src/App.js
@@ -12,6 +12,7 @@ import About from "./Components/About";
 
 function App() {
   let initTodo;
+  const [selectedTodoForUpdate, setSelectedTodoForUpdate] = useState(null); 
   if(localStorage.getItem("todos") === null){
     initTodo = [];
   }
@@ -19,14 +20,41 @@ function App() {
     initTodo = JSON.parse(localStorage.getItem("todos"));
   }
   const onDelete = (item) =>{
-    console.log("I am in delete",item.sNo);
     //let index = toDoList.indexOf(item);
     //toDoList.splice(index,1);
+    if(item.status === true){
+      if(window.confirm("Want to delete not done task"))
+      settoDoList(toDoList.filter((e)=>{
+        return e!==item
+      }));
+    }
+    else{
+      settoDoList(toDoList.filter((e)=>{
+        return e!==item
+      }));
+    }
+
+    localStorage.setItem("todos", JSON.stringify(toDoList));
+  }
+
+  const onUpdate = (item) =>{
+    setSelectedTodoForUpdate(item);
     settoDoList(toDoList.filter((e)=>{
       return e!==item
     }));
+    
+  }
 
-    localStorage.setItem("todos", JSON.stringify(toDoList));
+  const Done = (item) =>{
+    //settoDoList(toDoList[item.sNo - 1].status) = false;
+    //toDoList[item.sNo - 1].status = false;
+    toDoList.map((workItem)=>{
+      if(workItem.sNo === item.sNo)
+      workItem.status = false;
+      
+      return workItem;
+    });
+    settoDoList([...toDoList]);
   }
 
   const addTodo= (Title, Desc)=>{
@@ -34,7 +62,8 @@ function App() {
     const myTodo = {
       sNo : sno,
       title: Title,
-      description: Desc
+      description: Desc,
+      status: true
     };
     settoDoList([...toDoList, myTodo]);
 
@@ -45,12 +74,13 @@ function App() {
     localStorage.setItem("todos", JSON.stringify(toDoList));
   }, [toDoList])
 
+  console.log(toDoList);
   return (
     <div className="App">
       <Router>
-        <Navbar Title="ToDo List" SearchBar = {true} />
+        <Navbar Title="Work List" SearchBar = {true} />
         <Routes>
-          <Route exact path="/" element={[<AddTodo addTodo={addTodo} />,<Todos ToDoList ={toDoList} onDelete={onDelete} />]} ></Route>
+          <Route exact path="/" element={[<AddTodo addTodo={addTodo} selectedTodoForUpdate={selectedTodoForUpdate} />,<Todos ToDoList ={toDoList} onDelete={onDelete} onUpdate={onUpdate} Done={Done} />]} ></Route>
           <Route exact path="/About" element={<About/>} >
           </Route>
         </Routes>
